Added `JsonConverter` attribute, so System.Text.Json serialization is seamless
